Added Module Manifest and updated Functions
Updated functions to target Physical adapters only. This does impose a
limitation for the time being but the intention is to use this to rename
the Hardware NIC's before any Virtual NIC's exist. Happy to take
suggestions and contributions on how to work through Virtual NIC's also
but the issue there is that they don't always have a unique MAC address.

Added a Module Manifest.